+++ packages/contracts-rfq/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.5.7](https://github.com/synapsecns/sanguine/compare/@synapsecns/contracts-rfq@0.5.6...@synapsecns/contracts-rfq@0.5.7) (2024-09-26)
+
+**Note:** Version bump only for package @synapsecns/contracts-rfq
+
+
+
+
+
+## [0.5.6](https://github.com/synapsecns/sanguine/compare/@synapsecns/contracts-rfq@0.5.5...@synapsecns/contracts-rfq@0.5.6) (2024-09-26)
+
+**Note:** Version bump only for package @synapsecns/contracts-rfq
+
+
+
+
+
+## [0.5.5](https://github.com/synapsecns/sanguine/compare/@synapsecns/contracts-rfq@0.5.4...@synapsecns/contracts-rfq@0.5.5) (2024-09-26)
+
+**Note:** Version bump only for package @synapsecns/contracts-rfq
+
+
+
+
+
+## [0.5.4](https://github.com/synapsecns/sanguine/compare/@synapsecns/contracts-rfq@0.5.3...@synapsecns/contracts-rfq@0.5.4) (2024-09-26)
+
+**Note:** Version bump only for package @synapsecns/contracts-rfq
+
+
+
+
+
+## [0.5.3](https://github.com/synapsecns/sanguine/compare/@synapsecns/contracts-rfq@0.5.2...@synapsecns/contracts-rfq@0.5.3) (2024-09-26)
+
+**Note:** Version bump only for package @synapsecns/contracts-rfq
+
+
+
+
+
 ## [0.5.2](https://github.com/synapsecns/sanguine/compare/@synapsecns/contracts-rfq@0.5.1...@synapsecns/contracts-rfq@0.5.2) (2024-09-25)
 
 **Note:** Version bump only for package @synapsecns/contracts-rfq